# src/main/resources/APIDOC.YAML
openapi: 3.0.1
info:
  title: Constructa OpenApi specification
  description: OpenApi documentation for Constructa BE
  contact:
    name: Ondřej Bureš
    url: 'https://digitalnoodles.cz/'
    email: bureson2@fel.cvut.cz
  version: '1.0'
servers:
  - url: 'http://localhost:8080/'
    description: Local ENV
security:
  - bearerAuth: []
paths:
...
  /api/v1/users:
    get:
      tags:
        - Users
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Users
      operationId: updateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
      security:
        - bearerAuth: []
  /api/v1/tasks:
    get:
      tags:
        - Tasks
      operationId: getTasks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Tasks
      operationId: editTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskDTO'
      security:
        - bearerAuth: []
    post:
      tags:
        - Tasks
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskDTO'
      security:
        - bearerAuth: []
  /api/v1/projects:
    get:
      tags:
        - Projects
      operationId: getProjects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectDTO'
      security:
        - bearerAuth: []
    put:
      tags:
        - Projects
      operationId: editProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProjectDTO'
      security:
        - bearerAuth: []
    post:
      tags:
        - Projects
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProjectDTO'
      security:
        - bearerAuth: []
...
  /api/v1/tasks/state:
    post:
      tags:
        - Tasks
      operationId: editTaskState
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TaskDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/work-reports/user/{userId}':
    get:
      tags:
        - Work reports
      operationId: getUsersWorkReports
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkReportDTO'
      security:
        - bearerAuth: []
...
  '/api/v1/users/{userId}':
    get:
      tags:
        - Users
      operationId: getUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Users
      operationId: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
  /api/v1/users/search:
    get:
      tags:
        - Users
      operationId: getInputUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserInputDTO'
      security:
        - bearerAuth: []
  '/api/v1/tasks/{taskId}':
    get:
      tags:
        - Tasks
      operationId: getTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Tasks
      operationId: deleteTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
  /api/v1/tasks/my:
    get:
      tags:
        - Tasks
      operationId: getUserTasks
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskDTO'
      security:
        - bearerAuth: []
  '/api/v1/projects/{projectId}':
    get:
      tags:
        - Projects
      operationId: getProjects_1
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDTO'
      security:
        - bearerAuth: []
    delete:
      tags:
        - Projects
      operationId: deleteProject
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
      security:
        - bearerAuth: []
...
  '/api/v1/construction-reports/project/{projectId}':
    get:
      tags:
        - Construction reports
      operationId: getConstructionReportsByProjectId
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConstructionReportDTO'
      security:
        - bearerAuth: []
...
components:
  schemas:
...
    LocationDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        name:
          type: string
        country:
          type: string
        postCode:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        titleBeforeName:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        titleAfterName:
          type: string
        bankAccount:
          type: string
        dateOfAcceptance:
          type: string
          format: date-time
        dateOfBirth:
          type: string
          format: date-time
        birthId:
          type: string
        hourRate:
          type: integer
          format: int32
        monthSalary:
          type: integer
          format: int32
        address:
          $ref: '#/components/schemas/LocationDTO'
    WorkReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        minutes:
          type: integer
          format: int32
        reportingEmployee:
          $ref: '#/components/schemas/UserDTO'
        location:
          $ref: '#/components/schemas/LocationDTO'
        type:
          type: string
...
    UserRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        titleBeforeName:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        titleAfterName:
          type: string
        bankAccount:
          type: string
        email:
          type: string
        password:
          type: string
        dateOfBirth:
          type: string
          format: date-time
        birthId:
          type: string
        hourRate:
          type: integer
          format: int32
        monthSalary:
          type: integer
          format: int32
        phone:
          type: string
        role:
          type: string
          enum:
            - ROLE_ADMIN
            - ROLE_MANAGER
            - ROLE_REPORTER
            - ROLE_WAREHOUSE_MANAGER
            - ROLE_MECHANICAL_ENGINEER
            - ROLE_CONSTRUCTION_MANAGER
            - ROLE_DRIVER
            - ROLE_WORKER
            - ROLE_EMPLOYEE
            - ROLE_EXTERNALIST
            - ROLE_FIRED
        country:
          type: string
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        postCode:
          type: string
    TaskRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        locationName:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        state:
          type: string
        userId:
          type: integer
          format: int64
    TaskDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        dateOfCreation:
          type: string
          format: date-time
        name:
          type: string
        description:
          type: string
        timeFrom:
          type: string
          format: date-time
        timeTo:
          type: string
          format: date-time
        assignee:
          $ref: '#/components/schemas/UserDTO'
        author:
          $ref: '#/components/schemas/UserDTO'
        locationName:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        state:
          type: string
    ProjectRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        buldingFacility:
          type: string
        state:
          type: string
          enum:
            - FINISHED
            - IN_PREPARATION
            - IN_REALIZATION
        userId:
          type: integer
          format: int64
        startedAt:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        country:
          type: string
        city:
          type: string
        street:
          type: string
        descriptiveNumber:
          type: string
        postCode:
          type: string
    ProjectDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        buldingFacility:
          type: string
        state:
          type: string
        projectManager:
          $ref: '#/components/schemas/UserDTO'
        projectAddress:
          $ref: '#/components/schemas/LocationDTO'
        startedAt:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
...
    ConstructionReportDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        taskName:
          type: string
        note:
          type: string
        weather:
          type: string
        date:
          type: string
          format: date-time
        executor:
          $ref: '#/components/schemas/UserDTO'
        state:
          type: string
...
    UserInputDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        firstname:
          type: string
        lastname:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT auth description
      in: header
      scheme: bearer
      bearerFormat: JWT